// src/app/shared/model/shopping-address.model.ts
export class ShoppingAddress
{
    selectedCountryLinkId:number=0;
    selectedCountry: string='';
    selectedStateLinkId:number=0;
    selectedState: string='';
    selectedCityLinkId: number=0;
    selectedCity: string='';
    pincode: number=0;
    address: string='';

    flagCode: string='in';  
}